requirements.txt
pandas
numpy
scipy
scikit-learn
setuptools
dill
matplotlib
seaborn
python-dotenv
-e .